Fix #5713: Emit `used-before-assignment` instead of `undefined-variable` when accessing unused type annotations
