+++ pylint/checkers/variables.py
@@ -1543,7 +1543,10 @@ class VariablesChecker(BaseChecker):
                     )
 
         elif self._is_only_type_assignment(node, defstmt):
-            self.add_message("undefined-variable", args=node.name, node=node)
+            if node.scope().locals.get(node.name):
+                self.add_message("used-before-assignment", args=node.name, node=node)
+            else:
+                self.add_message("undefined-variable", args=node.name, node=node)
             return (VariableVisitConsumerAction.CONSUME, found_nodes)
 
         elif isinstance(defstmt, nodes.ClassDef):
