+++ tests/functional/u/use/used_before_assignment_type_annotations.txt
@@ -1,0 +1,3 @@
+used-before-assignment:15:10:15:18:only_type_assignment:Using variable 'variable' before assignment:UNDEFINED
+used-before-assignment:28:10:28:18:value_assignment_after_access:Using variable 'variable' before assignment:UNDEFINED
+undefined-variable:62:14:62:17:decorator_returning_incorrect_function.wrapper_with_type_and_no_value:Undefined variable 'var':UNDEFINED
